[refactor] Better import order by using __all__
