+++ python/taichi/core/__init__.py
@@ -1,9 +1,9 @@
-from .util import ti_core, build, format, load_module, start_memory_monitoring, \
-  is_release, package_root
+from .util import *
+from .settings import *
 from .unit import unit
 
 ti_core.build = build
 ti_core.format = format
 ti_core.load_module = load_module
 
-__all__ = [s for s in dir() if not s.startswith('_')] + ['settings']
+__all__ = [s for s in dir() if not s.startswith('_')]
