+++ python/taichi/core/util.py
@@ -6,8 +6,7 @@ import ctypes
 import warnings
 from pathlib import Path
 from colorama import Fore, Back, Style
-from taichi.misc.settings import get_output_directory, get_build_directory, get_bin_directory, get_repo_directory, get_runtime_directory
-from taichi.misc.util import get_os_name, get_unique_task_id
+from .settings import *
 
 if sys.version_info[0] < 3 or sys.version_info[1] <= 5:
     print("\nPlease restart with Python 3.6+\n")
@@ -488,3 +487,15 @@ def _print_taichi_header():
 
 
 _print_taichi_header()
+
+
+__all__ = [
+    'ti_core',
+    'build',
+    'format',
+    'load_module',
+    'start_memory_monitoring',
+    'is_release',
+    'package_root',
+    'require_version',
+    ]
