+++ python/taichi/misc/__init__.py
@@ -1,0 +1,8 @@
+from .util import *
+from .gui import *
+from .image import *
+from .test import *
+from .error import *
+from .task import Task
+
+__all__ = [s for s in dir() if not s.startswith('_')]
