+++ python/taichi/misc/gui.py
@@ -1,6 +1,6 @@
 import numbers
 import numpy as np
-from taichi import ti_core
+from taichi.core import ti_core
 
 
 class GUI:
@@ -342,3 +342,9 @@ class GUI:
 def rgb_to_hex(c):
     to255 = lambda x: min(255, max(0, int(x * 255)))
     return 65536 * to255(c[0]) + 256 * to255(c[1]) + to255(c[2])
+
+
+__all__ = [
+    'GUI',
+    'rgb_to_hex',
+    ]
