+++ python/taichi/tools/__init__.py
@@ -1,0 +1,4 @@
+from .video import VideoManager
+from .np2ply import PLYWriter
+
+__all__ = [s for s in dir() if not s.startswith('_')]
