+++ python/taichi/tools/video.py
@@ -1,4 +1,4 @@
-from taichi.misc.settings import get_os_name
+from taichi.core.settings import get_os_name
 from taichi.misc.image import imwrite
 
 import os
